refactor: Update `FeatureChip` to apply same optional removability feature that is on other data chips

+++ src/components/data-chips/feature-chip.svelte
@@ -8,27 +8,29 @@
 
 	type Props = {
 		feature: string
-		onRemove: (feature: string) => void
+		isRemoveable?: boolean
+		onRemove?: (feature: string) => void
 	}
 
-	const { feature, onRemove }: Props = $props()
-
-	const removeLabel = $derived(getDictionaryOf(LocaleContext.context.value!).common.dataChips.featureChip.labelRemove)
+	const { feature, isRemoveable = false, onRemove }: Props = $props()
+	const removeLabel = $derived(getDictionaryOf(LocaleContext.context.value!).common.dataChips.feature.ariaLabelRemove)
 </script>
 
-<li class="flex w-min items-center justify-start">
-	<div class="rounded-full px-4 py-2" style="background-color: {hslStrToCss(generateColorFrom(feature))}">
-		<p class="font-serif text-xl font-bold whitespace-nowrap text-obverse">
+<div class="group flex w-min items-center">
+	<div class="rounded-full px-2" style="background-color: {hslStrToCss(generateColorFrom(feature))}">
+		<p class="font-serif font-bold whitespace-nowrap text-obverse">
 			{feature}
 		</p>
 	</div>
-	<SlabButton
-		variant="text"
-		hasAccent={true}
-		fit="square"
-		ariaLabel={removeLabel}
-		size="normal"
-		onClick={() => onRemove(feature)}>
-		<Icon icon="Cancel" size="normal" />
-	</SlabButton>
-</li>
+	{#if isRemoveable && onRemove}
+		<SlabButton
+			variant="text"
+			hasAccent={true}
+			fit="square"
+			ariaLabel={removeLabel}
+			size="small"
+			onClick={() => onRemove(feature)}>
+			<Icon icon="Cancel" size="small" />
+		</SlabButton>
+	{/if}
+</div>
